# contracts/http/media-api/openapi.candidate.yaml
openapi: 3.1.0
info:
  title: media-api
  version: 3.2.0
  x-lifecycle: STABLE
  x-contract-id: media-api
  x-composition-version-range-grammar: composition-version-range-v1
  x-composition-version-range-grammar-document: docs/api/composition-version-range.md
  description: >
    VERSION_COMPATIBILITY_GOVERNANCE_FOUNDATION_V1 baseline. The OpenAPI document
    is authority for the EXTERNAL HTTP TRANSPORT CONTRACT only — it is NOT Media
    domain schema, Timeline canonical schema, capability semantic authority or
    revision authority (TRANSPORT_SCHEMA_IS_A_PROJECTION_NOT_DOMAIN_AUTHORITY_V1).
servers:
  - url: /api
paths:
  /composition/admissions:
    post:
      operationId: admitComposition
      summary: Admit a published Composition using server-owned authority facts
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CompositionAdmissionRequest'}}}}
      responses: {'200': {description: Durable admission decision, content: {application/json: {schema: {$ref: '#/components/schemas/CompositionAdmissionResponse'}}}}, '403': {description: Scope denied}, '409': {description: Idempotency conflict}, '422': {description: Invalid admission selectors or published composition}}
  /composition/capabilities:
    get:
      operationId: listCapabilityAvailability
      summary: Read provider-neutral capability availability
      responses:
        '200': {description: Capability availability catalog}
  /composition/workflows/drafts:
    post:
      operationId: saveTemplateWorkflowDraft
      summary: Save a Template Workflow draft
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TemplateWorkflow'}}}}
      responses: {'200': {description: Draft saved}}
  /composition/workflows/validate:
    post:
      operationId: validateTemplateWorkflow
      summary: Validate a Template Workflow without execution
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TemplateWorkflow'}}}}
      responses: {'200': {description: Validation result}}
  /composition/workflows/drafts/{id}:
    get:
      operationId: getTemplateWorkflowDraft
      summary: Get a workspace-owned workflow draft
      parameters: [{name: id, in: path, required: true, schema: {type: string}}, {name: workspaceId, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Draft}}
    put:
      operationId: updateTemplateWorkflowDraft
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TemplateWorkflow'}}}}
      responses: {'200': {description: Draft updated}}
  /composition/workflows/drafts/{id}/publish:
    post:
      operationId: publishTemplateWorkflow
      summary: Publish a validated workflow draft
      parameters: [{name: id, in: path, required: true, schema: {type: string}}, {name: workspaceId, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Published immutable version}}
  /composition/workflows/versions/{id}/{version}:
    get:
      operationId: getTemplateWorkflowVersion
      summary: Get an immutable workflow version
      parameters: [{name: id, in: path, required: true, schema: {type: string}}, {name: version, in: path, required: true, schema: {type: string}}, {name: workspaceId, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Immutable version}}
  /composition/applications/drafts:
    post:
      operationId: saveApplicationDraft
      summary: Save an Application draft
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Application'}}}}
      responses: {'200': {description: Draft saved}}
  /composition/applications/validate:
    post:
      operationId: validateApplication
      summary: Validate an Application without execution
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Application'}}}}
      responses: {'200': {description: Validation result}}
  /composition/applications/drafts/{id}:
    get:
      operationId: getApplicationDraft
      summary: Get an application draft
      parameters: [{name: id, in: path, required: true, schema: {type: string}}, {name: workspaceId, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Draft}}
    put:
      operationId: updateApplicationDraft
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Application'}}}}
      responses: {'200': {description: Draft updated}}
  /composition/applications/drafts/{id}/publish:
    post:
      operationId: publishApplication
      summary: Publish a validated application draft
      parameters: [{name: id, in: path, required: true, schema: {type: string}}, {name: workspaceId, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Published immutable version}}
  /composition/applications/versions/{id}/{version}:
    get:
      operationId: getApplicationVersion
      summary: Get an immutable application version
      parameters: [{name: id, in: path, required: true, schema: {type: string}}, {name: version, in: path, required: true, schema: {type: string}}, {name: workspaceId, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Immutable version}}
  /tenants/{tenantId}/projects/{projectId}/thumbnails:
    post:
      operationId: submitMediaThumbnail
      summary: Submit a durable timestamped thumbnail extraction
      parameters:
        - {name: tenantId, in: path, required: true, schema: {type: string}}
        - {name: projectId, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ThumbnailRequest'}
      responses:
        '200':
          description: Durable thumbnail task
          content: {application/json: {schema: {$ref: '#/components/schemas/ThumbnailTask'}}}
  /tenants/{tenantId}/projects/{projectId}/thumbnails/{taskId}:
    get:
      operationId: getMediaThumbnailTask
      summary: Get thumbnail task status
      parameters:
        - {name: tenantId, in: path, required: true, schema: {type: string}}
        - {name: projectId, in: path, required: true, schema: {type: string}}
        - {name: taskId, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Task status, content: {application/json: {schema: {$ref: '#/components/schemas/ThumbnailTask'}}}}
  /tenants/{tenantId}/projects/{projectId}/thumbnails/{taskId}/cancel:
    post:
      operationId: cancelMediaThumbnailTask
      summary: Cancel an admitted or running thumbnail task
      parameters:
        - {name: tenantId, in: path, required: true, schema: {type: string}}
        - {name: projectId, in: path, required: true, schema: {type: string}}
        - {name: taskId, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Cancelled task status, content: {application/json: {schema: {$ref: '#/components/schemas/ThumbnailTask'}}}}
  /tenants/{tenantId}/projects/{projectId}/thumbnails/{taskId}/image:
    get:
      operationId: getMediaThumbnailImage
      summary: Retrieve an authorized committed thumbnail image
      parameters:
        - {name: tenantId, in: path, required: true, schema: {type: string}}
        - {name: projectId, in: path, required: true, schema: {type: string}}
        - {name: taskId, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Committed image, content: {image/jpeg: {}, image/png: {}}}
  /render/jobs/{jobId}:
    get:
      operationId: getRenderJob
      summary: Query a render job (transport projection of render execution state)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Render job projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderJob"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /timeline/revisions/{revisionId}:
    get:
      operationId: getTimelineRevision
      summary: Fetch an immutable Timeline revision (composition projection)
      parameters:
        - name: revisionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Timeline revision projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineRevision"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /media/assets/{assetId}:
    head:
      operationId: headMediaAsset
      summary: Media asset existence probe (additive, non-breaking)
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Exists
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
    get:
      operationId: getMediaAsset
      summary: Media asset projection (media-domain logical source)
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Media asset projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAsset"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /composition/scope:
    get:
      operationId: resolveCompositionScope
      summary: Resolve the authenticated workspace scope
      parameters: [{name: workspaceId, in: query, required: false, schema: {type: string}}]
      responses: {'200': {description: Verified tenant/workspace scope}, '401': {description: Unauthenticated}, '403': {description: Ineligible membership}}
  /composition/workflows/publish-readiness:
    post:
      operationId: checkTemplateWorkflowPublishReadiness
      summary: Check workflow publish readiness without execution
      responses: {'200': {description: Typed validation result}}
  /composition/workflows/versions/{id}/compare:
    get:
      operationId: compareTemplateWorkflowVersions
      summary: Compare immutable workflow versions
      parameters: [{name: id, in: path, required: true, schema: {type: string}}, {name: workspaceId, in: query, required: true, schema: {type: string}}, {name: left, in: query, required: true, schema: {type: string}}, {name: right, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Version comparison}}
  /composition/applications/publish-readiness:
    post:
      operationId: checkApplicationPublishReadiness
      summary: Check application publish readiness without execution
      responses: {'200': {description: Typed validation result}}
  /composition/applications/versions/{id}/compare:
    get:
      operationId: compareApplicationVersions
      summary: Compare immutable application versions
      parameters: [{name: id, in: path, required: true, schema: {type: string}}, {name: workspaceId, in: query, required: true, schema: {type: string}}, {name: left, in: query, required: true, schema: {type: string}}, {name: right, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Version comparison}}
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    TemplateWorkflow:
      type: object
      required: [id, version, name, steps, bindings, parameters, requiredCapabilities, executionModes, requiredAssets, outputs, entry, estimate, reliability, lifecycle, tenantId, workspaceId, revision]
      properties:
        id: {type: string}
        version: {type: string}
        name: {type: string}
        steps: {type: array, items: {$ref: '#/components/schemas/WorkflowStep'}}
        bindings: {type: array, items: {$ref: '#/components/schemas/Binding'}}
        parameters: {type: array, items: {$ref: '#/components/schemas/Parameter'}}
        requiredCapabilities: {type: array, items: {type: string}}
        executionModes: {type: array, items: {type: string, enum: [SYNCHRONOUS, ASYNCHRONOUS, BATCH]}}
        requiredAssets: {type: array, items: {type: string}}
        entitlements: {type: array, items: {$ref: '#/components/schemas/EntitlementRequirement'}}
        outputs: {type: array, items: {$ref: '#/components/schemas/WorkflowOutput'}}
        entry: {$ref: '#/components/schemas/WorkflowEntry'}
        estimate: {$ref: '#/components/schemas/CostEstimate'}
        reliability: {$ref: '#/components/schemas/Reliability'}
        lifecycle: {type: string, enum: [DRAFT, PUBLISHED]}
        tenantId: {type: string}
        workspaceId: {type: string}
        revision: {type: integer, format: int64}
    Application:
      type: object
      required: [id, version, displayName, workflowIds]
      properties:
        id: {type: string}
        version: {type: string}
        displayName: {type: string}
        workflowIds: {type: array, items: {type: string}}
        requiredCapabilities: {type: array, items: {type: string}}
        requiredAssets: {type: array, items: {type: string}}
        entitlements: {type: array, items: {type: string}}
        executionModes: {type: array, items: {type: string, enum: [SYNCHRONOUS, ASYNCHRONOUS, BATCH]}}
        lifecycle: {type: string, enum: [DRAFT, PUBLISHED]}
        tenantId: {type: string}
        workspaceId: {type: string}
        revision: {type: integer, format: int64}
    EntitlementRequirement:
      type: object
      required: [key, version]
      properties:
        key: {type: string}
        version: {type: string}
    CompositionAdmissionRequest:
      type: object
      required: [workspaceSelection, compositionId, publishedVersion, idempotencyKey, requestHash, cancellationPolicy, retryPolicy]
      properties:
        workspaceSelection: {type: string}
        compositionId: {type: string}
        publishedVersion: {type: string}
        idempotencyKey: {type: string}
        requestHash: {type: string}
        cancellationPolicy: {type: string}
        retryPolicy: {type: string}
        parameters: {type: object, additionalProperties: true}
    CompositionAdmissionResponse:
      type: object
      required: [executionId, ownershipGeneration, state, newlyAdmitted, compositionId, version, revision, planFingerprint]
      properties:
        executionId: {type: string}
        ownershipGeneration: {type: integer, format: int64}
        state: {type: string}
        newlyAdmitted: {type: boolean}
        compositionId: {type: string}
        version: {type: string}
        revision: {type: integer, format: int64}
        planFingerprint: {type: string}
    WorkflowStep:
      type: object
      required: [id, capabilityId, capabilityVersion, inputs, requiredAssets, alternatives]
      properties:
        id: {type: string}
        capabilityId: {type: string}
        capabilityVersion: {type: string}
        inputs: {type: object, additionalProperties: true}
        requiredAssets: {type: array, items: {type: string}}
        alternatives: {type: array, items: {type: string}}
        entitlements: {type: array, items: {$ref: '#/components/schemas/EntitlementRequirement'}}
    Binding:
      type: object
      required: [fromStep, fromOutput, toStep, toInput, type]
      properties:
        fromStep: {type: string}
        fromOutput: {type: string}
        toStep: {type: string}
        toInput: {type: string}
        type: {type: string}
    WorkflowEntry:
      type: object
      required: [name, contract, stepId, port]
      properties:
        name: {type: string}
        contract: {$ref: '#/components/schemas/ContractRef'}
        stepId: {type: string}
        port: {type: string}
    WorkflowOutput:
      type: object
      required: [name, type, stepId, port]
      properties:
        name: {type: string}
        type: {type: string}
        stepId: {type: string}
        port: {type: string}
    Parameter:
      type: object
      required: [name, type, required]
      properties:
        name: {type: string}
        type: {type: string, enum: [string, integer, number, boolean, object, array]}
        defaultValue: {}
        minimum: {type: number}
        maximum: {type: number}
        required: {type: boolean}
    CostEstimate:
      type: object
      required: [units, unit, quotaUnits]
      properties:
        units: {type: number}
        unit: {type: string}
        quotaUnits: {type: number}
    Reliability:
      type: object
      required: [cancellable, retryable, maxRetries]
      properties:
        cancellable: {type: boolean}
        retryable: {type: boolean}
        maxRetries: {type: integer, minimum: 0}
    ValidationIssue:
      type: object
      required: [code, path, message, severity, objectType, objectId, location]
      properties:
        code: {type: string}
        path: {type: string}
        message: {type: string}
        severity: {type: string, enum: [ERROR, WARNING]}
        objectType: {type: string}
        objectId: {type: string}
        location: {type: string}
    ValidationResult:
      type: object
      required: [ready, issues, snapshotId]
      properties:
        ready: {type: boolean}
        issues: {type: array, items: {$ref: '#/components/schemas/ValidationIssue'}}
        snapshotId: {type: string}
    ThumbnailRequest:
      type: object
      required: [sourceAssetId, timestampSeconds, idempotencyKey]
      properties:
        sourceAssetId: {type: string}
        timestampSeconds: {type: number, format: double, minimum: 0}
        imageFormat: {type: string, enum: [jpeg, png], default: jpeg}
        width: {type: integer, minimum: 16, maximum: 4096}
        quality: {type: integer, minimum: 1, maximum: 100}
        idempotencyKey: {type: string, minLength: 1}
    ThumbnailTask:
      type: object
      required: [taskId, status]
      properties:
        taskId: {type: string}
        status: {type: string, enum: [ADMITTED, RUNNING, COMMITTING, COMPLETED, FAILED, CANCELLED]}
        artifactId: {type: string, nullable: true}
        failureCode: {type: string, nullable: true}
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
      x-error-shape: consistent
    RenderJob:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          type: string
        status:
          type: string
        revisionId:
          type: string
    TimelineRevision:
      type: object
      required: [revisionId, schemaVersion]
      properties:
        revisionId:
          type: string
        schemaVersion:
          type: string
        tracks:
          type: array
          items:
            type: object
    MediaAsset:
      type: object
      required: [assetId]
      properties:
        assetId:
          type: string
        streamId:
          type: string
    CapabilityAvailability:
      type: object
      required: [capabilityId, version, input, output, availability]
      properties:
        capabilityId: {type: string}
        version: {type: string}
        input: {$ref: '#/components/schemas/ContractRef'}
        output: {$ref: '#/components/schemas/ContractRef'}
        assetTypes: {type: array, items: {type: string}}
        mediaTypes: {type: array, items: {type: string}}
        executionModes: {type: array, items: {type: string}}
        availability: {type: string}
        summary: {type: string}
    ContractRef:
      type: object
      required: [name, version]
      properties:
        name: {type: string}
        version: {type: string}